Add files via upload
I've developed a web application using Django, with SQLite3 as the primary database, seamlessly integrated with React. This project is designed to play a crucial role in emergency situations, significantly reducing response times for first aid and empowering doctors to remotely assess and predict patient health.

The application encompasses features such as timely reminders for patient tablet consumption and real-time monitoring of sickness status. These functionalities are pivotal in optimizing processes and enhancing overall operational efficiency. By leveraging the Django framework, the backend efficiently manages patient data, ensuring the reliability and security of critical information.

The integration of React on the frontend provides a dynamic and responsive user interface, fostering a seamless interaction experience. This cohesive combination of Django and React not only streamlines the workflow for medical practitioners but also contributes to a more efficient and effective response to patient emergencies.

In essence, this web application acts as a sophisticated tool, harnessing the power of Django and React, and utilizing SQLite3 for data storage. It empowers healthcare professionals to swiftly respond to emergencies, predict patient health remotely, and monitor vital aspects, ultimately improving the overall quality of patient care.